Rename `codec.py` → `transforms.py`
